# openapi.yaml
openapi: 3.0.0
info:
  title: Card Benefits Admin API
  description: |
    Admin management API for card and benefit administration.
    
    ## Rate Limiting
    All endpoints are rate limited per admin user:
    - List operations: 100-150 requests per minute
    - Create operations: 50 requests per minute
    - Update operations: 50 requests per minute
    - Delete operations: 30 requests per minute
    
    ## Pagination
    All list endpoints support pagination:
    - Default limit: 20-50 items per page (varies by endpoint)
    - Maximum limit: 100 items per page
    - Responses include pagination metadata (total, page, limit, totalPages, hasMore)
    
    ## Authentication
    All endpoints require admin role authentication via cookie (auth_token)
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:3000
    description: Development
...
paths:
  /api/admin/cards:
    get:
      tags:
        - Cards
      summary: List all cards
      description: Retrieve paginated list of master cards with filtering and sorting
      operationId: listCards
      x-rateLimit:
        limit: 100
        window: minute
        description: 100 requests per minute per admin user
      parameters:
        - name: page
          in: query
          description: Page number (default 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page (max 100, default 20)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: issuer
          in: query
          description: Filter by issuer (max 255 chars)
          schema:
            type: string
            maxLength: 255
        - name: search
          in: query
          description: Search in card name and issuer (max 255 chars)
          schema:
            type: string
            maxLength: 255
        - name: isActive
          in: query
          description: Filter by active status
          schema:
            type: boolean
        - name: sortBy
          in: query
          description: Sort by field
          schema:
            type: string
            enum:
              - issuer
              - cardName
              - displayOrder
              - updatedAt
            default: displayOrder
        - name: sortDirection
          in: query
          description: Sort direction
          schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
      responses:
        "200":
          description: Cards retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Card"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not authorized (admin required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - cookieAuth: []

    post:
      tags:
        - Cards
      summary: Create a new card
      description: Create a new master card (issuer + cardName must be unique)
      operationId: createCard
      x-rateLimit:
        limit: 50
        window: minute
        description: 50 requests per minute per admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - issuer
                - cardName
                - defaultAnnualFee
                - cardImageUrl
              properties:
                issuer:
                  type: string
                  maxLength: 100
                  example: "American Express"
                cardName:
                  type: string
                  maxLength: 200
                  example: "Gold Card"
                defaultAnnualFee:
                  type: integer
                  minimum: 0
                  example: 250
                cardImageUrl:
                  type: string
                  format: uri
                  example: "https://example.com/cards/amex-gold.png"
      responses:
        "201":
          description: Card created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/Card"
                  message:
                    type: string
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Duplicate card (issuer + cardName combination exists)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - cookieAuth: []
...
  /api/admin/users:
    get:
      tags:
        - Users
      summary: List users
      operationId: listUsers
      x-rateLimit:
        limit: 100
        window: minute
        description: 100 requests per minute per admin user
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          description: Search by email or name (max 255 chars)
          schema:
            type: string
            maxLength: 255
        - name: role
          in: query
          schema:
            type: string
            enum:
              - ADMIN
              - USER
      responses:
        "200":
          description: Users retrieved
      security:
        - cookieAuth: []
...
components:
  schemas:
    Card:
      type: object
      properties:
        id:
          type: string
        issuer:
          type: string
        cardName:
          type: string
        defaultAnnualFee:
          type: integer
        cardImageUrl:
          type: string
          format: uri
        displayOrder:
          type: integer
        isActive:
          type: boolean
        isArchived:
          type: boolean
        benefitCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        hasMore:
          type: boolean
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        code:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: HTTP cookie authentication
security:
  - cookieAuth: []